Implemented support for dual PWM on Raspberry
I still need to test how two concurrent PWMs behave.

It's likely that when enabling the second one, since we reset the clock
and, more important, we disable everything in the PWM CTL reg, the first
one will stop.

It's also likely that the clock must be shared between the two PWMs,
once it's set with the first signal, the second one can only change the
M/S parameters. The PWMOutputs are separate objects, this behaviour could
be hard to manage.