Single channel bit-pattern PWM
First implementation of the bit-pattern PWM signal generator.
It can be used to generate digital signals where the 0 and 1 are
represented by a signal with the same duration but different duty cycle.

The generator configures the PWM to operate in FIFO mode (that behaves
like a PWM in serial mode but takes the data from a list of uint32) and
uses the DMA controller to feed data to the PWM. This data is obtained
expanding each bit of the original binary data to a series of bit-patterns
of variable widthi, obtained identifying the smaller bit sequence able to
represent the values 0 and 1 as distinct pattern.

This needs some documentation to be configured effectively.

Only one of the two available channels are used at the moment, adding
support for the second channel is quite straightforward. When used in
2-chan configuration, both signals will have to use the same frequency
and will need to be configured and generated at the same moment since
they share the FIFO (sadly, the data for the two channels is interleaved).

Tested with a WS2812B led strip, driver coming soon.