Conversion from byte data to bit stream four times faster
This was the main bottleneck of the sendDataWithPattern method,
extremely slow. Now it's quite fast.

Instead of multiple nested for, with an inner for iteration for each bit
of the pattern, the block of memory containing the FIFO rows is now
modified directly applying the patterns on the right location of the
sequence of uint8s, even when patterns overlap two or more bytes.

With this improvement, the library is as fast as other C libraries that
generate signals with bit patterns via PWM (i.e. jgarff ws2812 lib).