Added first example for bit-pattern PWM: WS2812
Nothing fancy for now, a few simple effects (snake,random leds and other
patterns) for a 60 leds strip (a few hundreds of leds are supported).

Using the PWM pin 18.

Feel free to play with it, changing the number of leds in the sequence
or displaying something else.

Tested on a WS2812(or 2811?) strip and a WS2812B 8x8 matrix (something
similar to an unicornhat but cheaper).