Restore DMA Behavior
This change should restore the DMA behavior after the refactoring to extract the PHY out. When initializing the PHY for the first time.

Even though we are getting information for more than one DMA channel on initialization now, the offset cost of not having to memmap /dev/mem for every channel helps here, if you need more than one DMA channel.