Use UInt32 for Register Pointers
The GPIO base pointers used for the Raspberry Pi boards are UnsafeMutablePointer<UInt>, assuming that UInt == UInt32.

When compiling SwiftyGPIO for arm64, this makes the stride calculations wrong, as it will jump ’n * 8’ bytes instead of ’n * 4’ bytes.

This change converts the base pointers to be using UnsafeMutablePointer<UInt32> instead, and then deals with all the fallout from that change.