Replace gnu_dev_makedev with Swift implementation
On systems with an up to date libc, gnu_dev_makedev is not imported anymore
(check out bits/sysmacros.h to see how it changed), so for now I'm re-implementing it here.
Not the best solution since the underlying implementation could change or could
be even different depending on the system you are on.
Will look into how to fix this directly in Swift.